feat: google-auth 전략 네이밍 변경 및 회원가입 로직 수정
