+++ src/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { UsersModule } from '@src/users/users.module';
 import { AuthService } from './auth.service';
-import { GoogleOauthStrategy } from './google-oauth.strategy';
+import { GoogleAuthStrategy } from './google-auth.strategy';
 import { JwtStrategy } from './jwt.strategy';
 import { LocalStrategy } from './local.strategy';
 import { RefreshJwtStrategy } from './refresh-jwt.startegy';
@@ -27,7 +27,7 @@ import { RefreshJwtStrategy } from './refresh-jwt.startegy';
     LocalStrategy,
     JwtStrategy,
     RefreshJwtStrategy,
-    GoogleOauthStrategy,
+    GoogleAuthStrategy,
   ],
   exports: [AuthService],
 })
