--- src/auth/google-oauth.strategy.ts
@@ -1,22 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { PassportStrategy } from '@nestjs/passport';
-import { Strategy } from 'passport-google-oauth20';
-
-@Injectable()
-export class GoogleOauthStrategy extends PassportStrategy(Strategy, 'google') {
-  constructor(private configService: ConfigService) {
-    super({
-      clientID: configService.get<string>('GOOGLE_CLIENT_ID'),
-      clientSecret: configService.get<string>('GOOGLE_CLIENT_SECRET'),
-      callbackURL: configService.get<string>('GOOGLE_CALLBACK_URI'),
-      scope: ['email', 'profile'],
-    });
-  }
-  async validate(accessToken, refreshToken, profile) {
-    return {
-      name: profile.displayName,
-      email: profile.emails[0].value,
-    };
-  }
-}
